+++ src/ReaderMiddleware.ts
@@ -751,6 +751,25 @@ export const chainW: <R2, I, E2, A, B>(
 ) => <R1, E1>(ma: ReaderMiddleware<R1, I, I, E1, A>) => ReaderMiddleware<R1 & R2, I, I, E1 | E2, B> = chain as any
 
 /**
+ * Alias of `chainW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMap: {
+  <R2, I, A, E2, B>(f: (a: A) => ReaderMiddleware<R2, I, I, E2, B>): <R1, E1>(
+    ma: ReaderMiddleware<R1, I, I, E1, A>
+  ) => ReaderMiddleware<R1 & R2, I, I, E1 | E2, B>
+  <R1, I, E1, A, R2, E2, B>(
+    ma: ReaderMiddleware<R1, I, I, E1, A>,
+    f: (a: A) => ReaderMiddleware<R2, I, I, E2, B>
+  ): ReaderMiddleware<R1 & R2, I, I, E1 | E2, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1
+    ? (ma: ReaderMiddleware<any, any, any, any, any>) => _chain(ma, args[0] as any)
+    : _chain(args[0] as any, args[1] as any)) as any
+
+/**
  * Less strict version of [`flatten`](#flatten).
  *
  * @category combinators
@@ -779,6 +798,23 @@ export const flatten: <R, I, E, A>(
 export const ichain: <R, A, O, Z, E, B>(
   f: (a: A) => ReaderMiddleware<R, O, Z, E, B>
 ) => <I>(ma: ReaderMiddleware<R, I, O, E, A>) => ReaderMiddleware<R, I, Z, E, B> = ichainW
+
+/**
+ * Indexed version of [`flatMap`](#flatmap). Alias of `ichainW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const iflatMap: {
+  <R2, A, O, Z, E2, B>(f: (a: A) => ReaderMiddleware<R2, O, Z, E2, B>): <R1, I, E1>(
+    ma: ReaderMiddleware<R1, I, O, E1, A>
+  ) => ReaderMiddleware<R1 & R2, I, Z, E1 | E2, B>
+  <R1, I, O, Z, E1, A, R2, E2, B>(
+    ma: ReaderMiddleware<R1, I, O, E1, A>,
+    f: (a: A) => ReaderMiddleware<R2, O, Z, E2, B>
+  ): ReaderMiddleware<R1 & R2, I, Z, E1 | E2, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? ichainW(args[0] as any) : ichainW(args[1] as any)(args[0] as any)) as any
 
 /**
  * Less strict version of [`ichain`](#ichain).
@@ -1477,6 +1513,192 @@ export const chainFirstTaskOptionK: <E>(
 ) => <A, B>(
   f: (a: A) => TO.TaskOption<B>
 ) => <R, I>(ma: ReaderMiddleware<R, I, I, E, A>) => ReaderMiddleware<R, I, I, E, A> = chainFirstTaskOptionKW
+
+/**
+ * Alias of `chainMiddlewareKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapMiddleware: {
+  <I, A, E2, B>(f: (a: A) => M.Middleware<I, I, E2, B>): <R, E1>(
+    ma: ReaderMiddleware<R, I, I, E1, A>
+  ) => ReaderMiddleware<R, I, I, E1 | E2, B>
+  <R, I, E1, A, E2, B>(ma: ReaderMiddleware<R, I, I, E1, A>, f: (a: A) => M.Middleware<I, I, E2, B>): ReaderMiddleware<
+    R,
+    I,
+    I,
+    E1 | E2,
+    B
+  >
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainMiddlewareKW(args[0] as any) : chainMiddlewareKW(args[1] as any)(args[0] as any)) as any
+
+/**
+ * Alias of `chainEitherKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapEither: {
+  <A, E2, B>(f: (a: A) => E.Either<E2, B>): <R, I, E1>(
+    ma: ReaderMiddleware<R, I, I, E1, A>
+  ) => ReaderMiddleware<R, I, I, E1 | E2, B>
+  <R, I, E1, A, E2, B>(ma: ReaderMiddleware<R, I, I, E1, A>, f: (a: A) => E.Either<E2, B>): ReaderMiddleware<
+    R,
+    I,
+    I,
+    E1 | E2,
+    B
+  >
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainEitherKW(args[0] as any) : chainEitherKW(args[1] as any)(args[0] as any)) as any
+
+/**
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapOption: {
+  <A, E2, B>(f: (a: A) => O.Option<B>, onNone: (a: A) => E2): <R, I, E1>(
+    ma: ReaderMiddleware<R, I, I, E1, A>
+  ) => ReaderMiddleware<R, I, I, E1 | E2, B>
+  <R, I, E1, A, E2, B>(
+    ma: ReaderMiddleware<R, I, I, E1, A>,
+    f: (a: A) => O.Option<B>,
+    onNone: (a: A) => E2
+  ): ReaderMiddleware<R, I, I, E1 | E2, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 2
+    ? (ma: ReaderMiddleware<any, any, any, any, any>) => (flatMapOption as any)(ma, args[0], args[1])
+    : flatMapEither(args[0] as any, (a: any) =>
+        pipe(
+          (args[1] as any)(a),
+          E.fromOption(() => (args[2] as any)(a))
+        )
+      )) as any
+
+/**
+ * Alias of `chainIOK`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapIO: {
+  <A, B>(f: (a: A) => IO<B>): <R, I, E>(ma: ReaderMiddleware<R, I, I, E, A>) => ReaderMiddleware<R, I, I, E, B>
+  <R, I, E, A, B>(ma: ReaderMiddleware<R, I, I, E, A>, f: (a: A) => IO<B>): ReaderMiddleware<R, I, I, E, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainIOK(args[0] as any) : chainIOK(args[1] as any)(args[0] as any)) as any
+
+/**
+ * Alias of `chainTaskK`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapTask: {
+  <A, B>(f: (a: A) => Task<B>): <R, I, E>(ma: ReaderMiddleware<R, I, I, E, A>) => ReaderMiddleware<R, I, I, E, B>
+  <R, I, E, A, B>(ma: ReaderMiddleware<R, I, I, E, A>, f: (a: A) => Task<B>): ReaderMiddleware<R, I, I, E, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainTaskK(args[0] as any) : chainTaskK(args[1] as any)(args[0] as any)) as any
+
+/**
+ * Alias of `chainTaskEitherKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapTaskEither: {
+  <A, E2, B>(f: (a: A) => TE.TaskEither<E2, B>): <R, I, E1>(
+    ma: ReaderMiddleware<R, I, I, E1, A>
+  ) => ReaderMiddleware<R, I, I, E1 | E2, B>
+  <R, I, E1, A, E2, B>(ma: ReaderMiddleware<R, I, I, E1, A>, f: (a: A) => TE.TaskEither<E2, B>): ReaderMiddleware<
+    R,
+    I,
+    I,
+    E1 | E2,
+    B
+  >
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainTaskEitherKW(args[0] as any) : chainTaskEitherKW(args[1] as any)(args[0] as any)) as any
+
+/**
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapTaskOption: {
+  <A, E2, B>(f: (a: A) => TO.TaskOption<B>, onNone: (a: A) => E2): <R, I, E1>(
+    ma: ReaderMiddleware<R, I, I, E1, A>
+  ) => ReaderMiddleware<R, I, I, E1 | E2, B>
+  <R, I, E1, A, E2, B>(
+    ma: ReaderMiddleware<R, I, I, E1, A>,
+    f: (a: A) => TO.TaskOption<B>,
+    onNone: (a: A) => E2
+  ): ReaderMiddleware<R, I, I, E1 | E2, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 2
+    ? (ma: ReaderMiddleware<any, any, any, any, any>) => (flatMapTaskOption as any)(ma, args[0], args[1])
+    : flatMapTaskEither(args[0] as any, (a: any) =>
+        TE.fromTaskOption(() => (args[2] as any)(a))((args[1] as any)(a))
+      )) as any
+
+/**
+ * Alias of `chainReaderKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapReader: {
+  <A, R2, B>(f: (a: A) => Reader<R2, B>): <R1, I, E>(
+    ma: ReaderMiddleware<R1, I, I, E, A>
+  ) => ReaderMiddleware<R1 & R2, I, I, E, B>
+  <R1, I, E, A, R2, B>(ma: ReaderMiddleware<R1, I, I, E, A>, f: (a: A) => Reader<R2, B>): ReaderMiddleware<
+    R1 & R2,
+    I,
+    I,
+    E,
+    B
+  >
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainReaderKW(args[0] as any) : chainReaderKW(args[1] as any)(args[0] as any)) as any
+
+/**
+ * Alias of `chainReaderTaskKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapReaderTask: {
+  <A, R2, B>(f: (a: A) => ReaderTask<R2, B>): <R1, I, E>(
+    ma: ReaderMiddleware<R1, I, I, E, A>
+  ) => ReaderMiddleware<R1 & R2, I, I, E, B>
+  <R1, I, E, A, R2, B>(ma: ReaderMiddleware<R1, I, I, E, A>, f: (a: A) => ReaderTask<R2, B>): ReaderMiddleware<
+    R1 & R2,
+    I,
+    I,
+    E,
+    B
+  >
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1 ? chainReaderTaskKW(args[0] as any) : chainReaderTaskKW(args[1] as any)(args[0] as any)) as any
+
+/**
+ * Alias of `chainReaderTaskEitherKW`.
+ *
+ * @category sequencing
+ * @since 0.8.0
+ */
+export const flatMapReaderTaskEither: {
+  <A, R2, E2, B>(f: (a: A) => ReaderTaskEither<R2, E2, B>): <R1, I, E1>(
+    ma: ReaderMiddleware<R1, I, I, E1, A>
+  ) => ReaderMiddleware<R1 & R2, I, I, E1 | E2, B>
+  <R1, I, E1, A, R2, E2, B>(
+    ma: ReaderMiddleware<R1, I, I, E1, A>,
+    f: (a: A) => ReaderTaskEither<R2, E2, B>
+  ): ReaderMiddleware<R1 & R2, I, I, E1 | E2, B>
+} = /*#__PURE__*/ ((...args: ReadonlyArray<unknown>) =>
+  args.length === 1
+    ? chainReaderTaskEitherKW(args[0] as any)
+    : chainReaderTaskEitherKW(args[1] as any)(args[0] as any)) as any
 
 /**
  * Phantom type can't be infered properly, use [`bindTo`](#bindto) instead.